Rename `testing` module and feature to `dev` to avoid confusion with `test`
